chore: read assets from nist cumulus export
NIST used asset management software called 'Canto Cumulus' to build what they referred to as an
"organised database" of photos and videos. Cumulus supported export, and one of the FOIA requests
included two files which represented data exports from two Cumulus databases, one for photos and one
for videos. These are useful because they contain the tags that NIST applied to photos and videos,
which was essentially how they organised the content.

This first commit contains some code that reads the export file, though so far it was only tested on
the photos export. Right now it just reads the asset data and doesn't convert it into anything. This
should follow at some point later.