feat: provide `import-releases` command
This also provides a supplementary command, `download-torrents`, because the torrents are required
for information for the releases.

We use the `diesel` ORM to save the release information in a Postgres database. The information for
the releases is provided from a combination of static data that is encoded in the binary and data
from the torrent file, e.g., file count and size.