chore: partially implement image import
Provides the beginning of a command to import images into the database. The entries in the Cumulus
image export are read into a hash table, where each key is a hash of the file name and size of the
image. The idea would be, when we are importing the content of a release into the database, as we
iterate over each image, we hash its file name and size, then check if there is a key in the table
that matches that hash. If there is, we can then apply the metadata from the Cumulus entry to that
image; otherwise, we import the image without any metadata.