refactor: use sqlx for `import-releases` command
I started to find the `diesel` ORM tedious to work with, especially when used with many to many
relationships. The amount of time I spent trying to understand all the attribute usage didn't seem
worth it.

I found the `sqlx` crate, which seems to be much simpler and more suited to the needs I have for
this small application.

In addition, the commands for the UI are refactored into sub-commands and there is a simple
`releases ls` command implemented here.