feat: provide `images import` command
The Cumulus image export file is read and its information is stored in a table. Then we iterate over
each file in the release and check whether it is in the Cumulus export, based on the name and size
of the file. If it's in the export, we will save the image in the database along with the metadata
from the export; otherwise, we will just save the image.

When importing the image, we read its dimensions from the file. There are two different crates in
use for this, because the `image` crate had a problem reading some of the TIF files in release 16.
The fallback mechanism is using an ImageMagick crate, so obviously that will need to be installed on
the machine where the code runs.