feat: provide `images convert` command
Converts the Cumulus image dump to a CSV file based on fields I've determined to be of interest.
Having the data in a spreadsheet application can be very useful for a high level overview and for
finding things.

This also implements a `get` command which searches for an asset in the dump with a particular name.
This was provided to examine whether the asset parsing was correct.