refactor: store torrent content in database
In the `releases init` command, the torrent content now gets stored in the database. It's stored in
a separate table just because that makes it easier to query the `releases` table without excluding
the torrent content, which in some cases can be several megabytes of data.

As a result, a couple of commands were changed to use release IDs rather than torrent paths.