feat: provide `add-master` command
This command uses `dialoguer::Editor` to provide an interactive command that adds a master video
record to the database. It would be possible to extend the command later to provide arguments for
the values, but that probably wouldn't be very useful. In most cases, the description field will
encompass multiple lines.

The database schema for representing a master video was brought in line with the spreadsheet I had
been using for building the database.