feat: provide `master-videos edit` command
Provides the ability to edit a master video record.

Some adjustments were made to change the `save_master_video` into an 'upsert' function, and at the
same time, the function for saving a list was removed.

A new `master-videos` subcommand was created and the `add-master` command was also moved there.