chore: store release files in the database
Create a new `release_files` table that will store all the files associated with a release.

This will subsequently be used to associate videos and the files in a release. The `content` table
that exists at present will most likely be removed in favour of separate image and video tables.