feat: provide `videos add` command
Much like the `master-videos add` command, this also enables an editor to be used to create a video
record, although it also provides a `--path` argument that can process a populated template, which
is useful for testing.

All the necessary database schema is defined here for the video, including the associated
relationship tables.