chore: convert from single crate to workspace
I need to be able to establish a database connection to my SQLite YouTube database, but it turns out
`sqlx` doesn't yet support multiple database connections within the same crate. Therefore the single
crate turns into a workspace, with one crate mainly dealing with database interaction.

The arrangement of types in the database crate doesn't completely make sense, because things like
the Cumulus image processing is there, due to a coupling with the database. However, it will do for
now, and can be refactored later.