chore: restructure videos representation in db
This is an absolutely enormous commit which should have been broken down into several, but it was
difficult to do so.

It completely restructures the representation of master and video records in the database, after
discovering the previous one wasn't fit for purpose. The master record will group the same videos,
and a new `news_broadcasts` table will group many masters related to the same news broadcast. There
is also a new `event_timestamps` table that supports categorisation of events. Features of Postgres,
like custom types and arrays, are leveraged to reduce many-to-many relationship tables.