refactor: use forms rather than ad-hoc parsing
I don't know why it never occurred to me to do it this way before. It's a much improved way of
handling the user input and editing. Doesn't quite save as much code as I hoped, but it's definitely
a better approach. New database model types are easier to add.