feat: provide `nist tapes edit` command
We want the NIST tapes database to be a copy of what was provided by NIST (despite the fact that
it's a bit of a mess), but in addition, to be able to associate release files with the tape. So
editing the tape simply provides the ability to associate release files with it.