feat: `nist tapes ls` command groups tapes by video
The tapes are all related to a video record, and this can group them together, which seems to be a
useful way of working with the list. So the command now outputs the list in this form.

BREAKING CHANGE: all previous arguments for the command have been removed. They don't seem as useful
with the way the new list is structured, although perhaps they will be introduced again.