feat: provide `--sum` arg on `releases ls-extensions` cmd
Before this change, the `releases ls-extensions` command functioned the way it does now with the
`--sum` argument. The next commit will implement the new functionality for `ls-extensions`, which
will be to display the extensions individually for each release in the range.

BREAKING CHANGE: the `--sum` argument now needs to be used to get the previous behaviour.