feat: mark nist videos as missing with notes
Some of the videos and tapes that were listed in NIST's database did not have a corresponding
release in the FOIA request fulfilments. I wanted the ability to distinguish these entries on the
`nist tapes ls` command.

To that end, the `nist_videos` table has been extended with two additional columns that were not in
the original database: `is_missing` and `additional_notes`. The flag gives the user the ability to
mark a video missing, and an additional notes field can be used to provide an explanation or other
supplementary information about the missing video. There is already a `notes` field on the original
table, but we want to preserve those notes so we can see how NIST noted the entry.